Consolidate validator classes
