+++ phpstan.neon.dist
@@ -15,6 +15,7 @@ parameters:
 		- '#Call to an undefined method CodeIgniter\\Database\\BaseBuilder::[A-Za-z]+\(\)#'
 		- '#Call to an undefined static method Config\\Services::[A-Za-z]+\(\)#'
 		- '#Cannot access property [\$a-z_]+ on (array|object)#'
+		- '#Parameter \$user of method Myth\\Auth\\Authentication\\Passwords\\ValidatorInterface::check\(\) has typehint with deprecated class CodeIgniter\\Entity#'
 	universalObjectCratesClasses:
 		- CodeIgniter\Entity\Entity
 		- Faker\Generator
