+++ src/Authentication/Passwords/BaseValidator.php
@@ -4,12 +4,26 @@ namespace Myth\Auth\Authentication\Passwords;
 
 use Myth\Auth\Config\Auth as AuthConfig;
 
-class BaseValidator
+abstract class BaseValidator
 {
     /**
-     * @var AuthConfig
+     * @var AuthConfig|null
      */
     protected $config;
+
+    /**
+     * Error message.
+     *
+     * @var string
+     */
+    protected $error = '';
+
+    /**
+     * Suggestion message.
+     *
+     * @var string
+     */
+    protected $suggestion = '';
 
     /**
      * Allows for setting a config file on the Validator.
@@ -23,5 +37,24 @@ class BaseValidator
         $this->config = $config;
 
         return $this;
+    }
+
+    /**
+     * Returns the error string that should be displayed to the user.
+     */
+    public function error(): string
+    {
+        return $this->error;
+    }
+
+    /**
+     * Returns a suggestion that may be displayed to the user
+     * to help them choose a better password. The method is
+     * required, but a suggestion is optional. May return
+     * an empty string instead.
+     */
+    public function suggestion(): string
+    {
+        return $this->suggestion;
     }
 }
