+++ src/Authentication/Passwords/DictionaryValidator.php
@@ -13,16 +13,6 @@ use CodeIgniter\Entity\Entity;
 class DictionaryValidator extends BaseValidator implements ValidatorInterface
 {
     /**
-     * @var string
-     */
-    protected $error;
-
-    /**
-     * @var string
-     */
-    protected $suggestion;
-
-    /**
      * Checks the password against the words in the file and returns false
      * if a match is found. Returns true if no match is found.
      * If true is returned the password will be passed to next validator.
@@ -50,24 +40,5 @@ class DictionaryValidator extends BaseValidator implements ValidatorInterface
         fclose($fp);
 
         return true;
-    }
-
-    /**
-     * Returns the error string that should be displayed to the user.
-     */
-    public function error(): string
-    {
-        return $this->error ?? '';
-    }
-
-    /**
-     * Returns a suggestion that may be displayed to the user
-     * to help them choose a better password. The method is
-     * required, but a suggestion is optional. May return
-     * an empty string instead.
-     */
-    public function suggestion(): string
-    {
-        return $this->suggestion ?? '';
     }
 }
