+++ src/Authentication/Passwords/NothingPersonalValidator.php
@@ -12,16 +12,6 @@ use CodeIgniter\Entity\Entity;
 class NothingPersonalValidator extends BaseValidator implements ValidatorInterface
 {
     /**
-     * @var string
-     */
-    protected $error;
-
-    /**
-     * @var string
-     */
-    protected $suggestion;
-
-    /**
      * Returns true if $password contains no part of the username
      * or the user's email. Otherwise, it returns false.
      * If true is returned the password will be passed to next validator.
@@ -201,24 +191,5 @@ class NothingPersonalValidator extends BaseValidator implements ValidatorInterfa
         }
 
         return $parts;
-    }
-
-    /**
-     * Returns the error string that should be displayed to the user.
-     */
-    public function error(): string
-    {
-        return $this->error ?? '';
-    }
-
-    /**
-     * Returns a suggestion that may be displayed to the user
-     * to help them choose a better password. The method is
-     * required, but a suggestion is optional. May return
-     * an empty string instead.
-     */
-    public function suggestion(): string
-    {
-        return $this->suggestion ?? '';
     }
 }
