+++ src/Authentication/Passwords/PwnedValidator.php
@@ -2,7 +2,7 @@
 
 namespace Myth\Auth\Authentication\Passwords;
 
-use CodeIgniter\Entity;
+use CodeIgniter\Entity\Entity;
 use CodeIgniter\HTTP\Exceptions\HTTPException;
 use Myth\Auth\Exceptions\AuthException;
 
@@ -19,20 +19,6 @@ use Myth\Auth\Exceptions\AuthException;
  */
 class PwnedValidator extends BaseValidator implements ValidatorInterface
 {
-    /**
-     * Error message.
-     *
-     * @var string
-     */
-    protected $error;
-
-    /**
-     * Suggestion message.
-     *
-     * @var string
-     */
-    protected $suggestion;
-
     /**
      * Checks the password against the online database and
      * returns false if a match is found. Returns true if no match is found.
@@ -83,24 +69,5 @@ class PwnedValidator extends BaseValidator implements ValidatorInterface
         $this->suggestion = lang('Auth.suggestPasswordPwned', [$password]);
 
         return false;
-    }
-
-    /**
-     * Returns the error string that should be displayed to the user.
-     */
-    public function error(): string
-    {
-        return $this->error;
-    }
-
-    /**
-     * Returns a suggestion that may be displayed to the user
-     * to help them choose a better password. The method is
-     * required, but a suggestion is optional. May return
-     * an empty string instead.
-     */
-    public function suggestion(): string
-    {
-        return $this->suggestion;
     }
 }
